+++ services/user/internal/adapters/postgres/user_profile_repository.go
@@ -20,12 +20,18 @@ func NewUserProfileRepository(db *postgres.Client) repository.UserProfileReposit
 	return &userProfileRepository{db: db}
 }
 
-func (r *userProfileRepository) CreateUserProfile(ctx context.Context, userProfile *entity.UserProfile) error {
+func (r *userProfileRepository) CreateUserProfile(
+	ctx context.Context,
+	userProfile *entity.UserProfile) error {
+
 	return r.db.GormDB.WithContext(ctx).
 		Create(userProfile).Error
 }
 
-func (r *userProfileRepository) UpdateLastLogin(ctx context.Context, userID uuid.UUID) error {
+func (r *userProfileRepository) UpdateLastLogin(
+	ctx context.Context,
+	userID uuid.UUID) error {
+
 	now := time.Now().UTC()
 	return r.db.GormDB.WithContext(ctx).
 		Model(&entity.UserProfile{}).
@@ -36,7 +42,10 @@ func (r *userProfileRepository) UpdateLastLogin(ctx context.Context, userID uuid
 		}).Error
 }
 
-func (r *userProfileRepository) ProfileExists(ctx context.Context, userID uuid.UUID) (bool, error) {
+func (r *userProfileRepository) ProfileExists(
+	ctx context.Context,
+	userID uuid.UUID) (bool, error) {
+
 	var matched int64
 	err := r.db.GormDB.WithContext(ctx).
 		Model(&entity.UserProfile{}).
@@ -48,7 +57,10 @@ func (r *userProfileRepository) ProfileExists(ctx context.Context, userID uuid.U
 	return matched > 0, nil
 }
 
-func (r *userProfileRepository) GetProfileByUserID(ctx context.Context, userID uuid.UUID) (*entity.UserProfile, error) {
+func (r *userProfileRepository) GetProfileByUserID(
+	ctx context.Context,
+	userID uuid.UUID) (*entity.UserProfile, error) {
+
 	var profile entity.UserProfile
 	err := r.db.GormDB.WithContext(ctx).
 		Model(&entity.UserProfile{}).
@@ -63,14 +75,20 @@ func (r *userProfileRepository) GetProfileByUserID(ctx context.Context, userID u
 	return &profile, nil
 }
 
-func (r *userProfileRepository) UpdateUserProfile(ctx context.Context, userProfile *entity.UserProfile) error {
+func (r *userProfileRepository) UpdateUserProfile(
+	ctx context.Context,
+	userProfile *entity.UserProfile) error {
+
 	return r.db.GormDB.WithContext(ctx).
 		Model(&entity.UserProfile{}).
 		Where("user_id = ?", userProfile.UserID).
 		Updates(userProfile).Error
 }
 
-func (r *userProfileRepository) GetUserProfileByID(ctx context.Context, id uint) (*entity.UserProfile, error) {
+func (r *userProfileRepository) GetUserProfileByID(
+	ctx context.Context,
+	id int64) (*entity.UserProfile, error) {
+
 	var profile entity.UserProfile
 	err := r.db.GormDB.WithContext(ctx).
 		Model(&entity.UserProfile{}).
@@ -95,10 +113,10 @@ func (r *userProfileRepository) GetPotentialProfiles(
 	isUserBride bool) ([]*entity.UserProfile, int64, error) {
 
 	// 1) Build base query
-    baseQuery := r.db.GormDB.WithContext(ctx).
-        Model(&entity.UserProfile{}).
-        Where("user_id <> ?", userID).
-        Where("is_bride = ?", !isUserBride)
+	baseQuery := r.db.GormDB.WithContext(ctx).
+		Model(&entity.UserProfile{}).
+		Where("user_id <> ?", userID).
+		Where("is_bride = ?", !isUserBride)
 
 	if len(excludedIDs) > 0 {
 		baseQuery = baseQuery.Where("user_id NOT IN (?)", excludedIDs)
@@ -106,26 +124,26 @@ func (r *userProfileRepository) GetPotentialProfiles(
 
 	// Apply preferences filters if provided
 	if preferences != nil {
-        baseQuery = r.applyPreferencesFilters(baseQuery, preferences)
-    }
+		baseQuery = r.applyPreferencesFilters(baseQuery, preferences)
+	}
 
 	// 2) Count total matches
-    var total int64
-    if err := baseQuery.
-        Session(&gorm.Session{}). // clone so we don’t carry over Limit/Offset/etc.
-        Count(&total).Error; err != nil {
-        return nil, 0, err
-    }
+	var total int64
+	if err := baseQuery.
+		Session(&gorm.Session{}). // clone so we don’t carry over Limit/Offset/etc.
+		Count(&total).Error; err != nil {
+		return nil, 0, err
+	}
 
 	// 3) Fetch paginated results
-    var profiles []*entity.UserProfile
-    if err := baseQuery.
-        Order("last_login DESC").
-        Limit(limit).
-        Offset(offset).
-        Find(&profiles).Error; err != nil {
-        return nil, 0, err
-    }
+	var profiles []*entity.UserProfile
+	if err := baseQuery.
+		Order("last_login DESC").
+		Limit(limit).
+		Offset(offset).
+		Find(&profiles).Error; err != nil {
+		return nil, 0, err
+	}
 
 	return profiles, total, nil
 }
