+++ services/user/internal/domain/entity/user_image.go
@@ -8,7 +8,7 @@ import (
 )
 
 type UserImage struct {
-    ID           uint           `gorm:"primaryKey;autoIncrement"`
+    ID           int64           `gorm:"primaryKey;autoIncrement"`
     UserID       uuid.UUID      `gorm:"type:uuid;not null;index:idx_user_images_user_id"`
     ObjectKey    string         `gorm:"type:varchar(500);not null"`
     DisplayOrder int16          `gorm:"type:smallint;not null;check:display_order >= 1 AND display_order <= 3;index:uniq_active_img_order,unique"`
