+++ services/user/internal/handlers/grpc/v1/confirm_profile_photo_upload.go
@@ -37,7 +37,7 @@ func (h *UserHandler) ConfirmProfilePhotoUpload(
 		ctx, 
 		userIDUUID, 
 		req.ObjectKey.Value, 
-		req.FileSize.Value,
+		int64(req.FileSize.Value),
 	)
 	if err != nil {
 		if !appErrors.IsAppError(err) {
