+++ services/user/internal/handlers/grpc/v1/get_profile_photo_upload_url.go
@@ -45,6 +45,6 @@ func (h *UserHandler) GetProfilePhotoUploadURL(
 	return &userpbv1.GetProfilePhotoUploadURLResponse{
 		UploadUrl:        &wrapperspb.StringValue{Value: response.UploadURL},
 		ObjectKey:        &wrapperspb.StringValue{Value: response.ObjectKey},
-		ExpiresInSeconds: &wrapperspb.UInt32Value{Value: response.ExpiresInSeconds},
+		ExpiresInSeconds: &wrapperspb.Int32Value{Value: int32(response.ExpiresInSeconds)},
 	}, nil
 }
